move the stuff to runtime (but it breaks it!! need fix!!)
