+++ loader/aarch64.ld
@@ -8,6 +8,12 @@ PHDRS
     all PT_LOAD AT (LINK_ADDRESS);
 }
 
+
+//    text PT_LOAD FLAGS(5); /* RX */
+//    rodata PT_LOAD FLAGS(4); /* RO */
+//    data PT_LOAD FLAGS(6); /* RW */
+//    bss PT_LOAD FLAGS(6); /* RW */
+
 SECTIONS
 {
     . = LINK_ADDRESS;
@@ -17,20 +23,26 @@ SECTIONS
     .text :
     {
         _text = .;
+
         KEEP(*(.text.start))
-        *(.text*)
-        *(.text.*)
-        *(.rodata)
-        *(.rodata.*)
+        *(.text .text.*)
+
         _text_end = .;
+    } :all
+
+    .rodata :
+    {
+        *(.rodata .rodata.* .rodata..Lanon.*)
     } :all
 
     .data :
     {
         _data = .;
-        *(.data)
+        *(.data .data.*)
         *(.data.*)
+
         KEEP(*(.data.uart_addr))
+
         _data_end = .;
     } :all
 
@@ -43,6 +55,8 @@ SECTIONS
         . = ALIGN(4);
         _bss_end = .;
     } :all
+
+
 
     _loader_end = .;
 }
