+++ loader/src/aarch64/mmu.c
@@ -20,8 +20,23 @@ uint64_t aarch64_pt_ttbr0_el1;
 uint64_t aarch64_pt_ttbr1_el1;
 uint64_t aarch64_pt_ttbr0_el2;
 
+struct ret {
+    uint64_t a;
+    uint64_t b;
+    uint64_t c;
+};
+
+extern struct ret aarch64_setup_pagetables(uint64_t kernel_first_vaddr, uint64_t kernel_first_paddr, uint64_t page_tables_paddr_start);
+
 int arch_mmu_enable(int logical_cpu)
 {
+    puts("setup1\n");
+    struct ret x = aarch64_setup_pagetables(0, 0, 0);
+    aarch64_pt_ttbr0_el1 = x.a;
+    aarch64_pt_ttbr1_el1 = x.b;
+    aarch64_pt_ttbr0_el2 = x.c;
+    puts("setup\n");
+
     int r;
     enum el el;
     r = ensure_correct_el(logical_cpu);
